requirements.txt
# Core LangChain dependencies
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters

# Google AI and Vector Store
google-generativeai
faiss-cpu

# Web Scraping and PDF Processing
beautifulsoup4
requests
lxml
PyPDF2
bs4

# Data Processing and Visualization
numpy
matplotlib
tqdm

# Configuration and Utilities
python-dotenv
PyYAML
colorama
rich
argparse
tenacity

# Development Tools
pytest
black

# Additional AI Services
fal-client
langchain-groq
Panel
Console